# src/mdm.api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "行動裝置管理系統 API — 裝置管控、資產管理、借用流程、App 派發",
        "title": "MDM Server API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "使用者自行註冊（需等待管理員啟用）",
                "parameters": [
                    {
                        "description": "註冊資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.swagRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    },
                    "409": {
                        "description": "帳號已存在",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
...
        "/api/setup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "初始化系統（建立第一位管理員）",
                "parameters": [
                    {
                        "description": "管理員帳號資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.swagSetupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    },
                    "403": {
                        "description": "系統已初始化",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
...
        "/api/system-status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "系統初始化狀態",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagSystemStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
        "/api/users-list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "使用者列表",
                "responses": {
                    "200": {
                        "description": "{users: [...]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/controller.swagUserListItem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "更新 / 刪除使用者（需 sys_admin）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新欄位（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagUserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "更新 / 刪除使用者（需 sys_admin）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新欄位（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagUserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/ws-config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "WebSocket 設定",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagWSConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.swagError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "controller.swagOK": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "controller.swagRegisterReq": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "New User"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                },
                "username": {
                    "type": "string",
                    "example": "newuser"
                }
            }
        },
...
        "controller.swagSetupReq": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Admin"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "controller.swagSystemStatus": {
            "type": "object",
            "properties": {
                "initialized": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controller.swagUserListItem": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "controller.swagUserUpdateReq": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "controller.swagWSConfig": {
            "type": "object",
            "properties": {
                "backend_relay": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}